Update twitch.py
Added links for Twitch Kraken API (V5), and the Twitch API Reference Guide
Added check to see if the twitch stream was in-fact, online before handling the download clips. (this will cut back on a lot of 're-downloading' checks)
Added longer wait times between queries when streamer has 'gone offline' or was simply, not online.
The checks added are based off an old website code I used and may no longer work. I may have to re-invent this when I do my checks on Monday.  But I wanted this at least, partially in place while I was thinking about it.